Agents/requirements.txt
# requirements.txt
google-adk
python-dotenv
google-generativeai
tidb-vector
pymysql
pytz